extract vaultDirector kind
